fix endpoint orders
