+++ controllers/order.js
@@ -7,23 +7,37 @@ const UserDao = require('../dao/user');
 class OrderController {
   static PaymentCategory = PaymentConstants.category;
   static async createOrder(req, res) {
+    let paymentLink;
     const { category, amount, userEmail } = req.body;
     const data = await Token.decryptJWT(req);
-    console.log(data);
     if (category === OrderController.PaymentCategory.recurrent) {
       try {
-        const paymentLink = await PaymentService.createSubscription(amount);
-        const user = await UserDao.findOneUser({ email: data.email });
-        await OrderDao.createOrder({
-          category,
+        paymentLink = await PaymentService.createSubscription(
           amount,
-          paymentLink,
-          user_id: user.id,
-        });
-        res.status(200).json(paymentLink);
+          'ARS',
+          userEmail
+        );
       } catch (error) {
         res.status(402).send(error.message);
       }
+    } else if (category === OrderController.PaymentCategory.single) {
+      try {
+        paymentLink = await PaymentService.createPayment(userEmail, amount);
+      } catch (error) {
+        res.status(402).send(error.message);
+      }
+    }
+    try {
+      const user = await UserDao.findOneUser({ email: data.email });
+      await OrderDao.createOrder({
+        category,
+        amount,
+        paymentLink,
+        user_id: user.id,
+      });
+      res.status(200).json(paymentLink);
+    } catch (error) {
+      res.status(402).send(error.message);
     }
   }
 }
