+++ dao/order.js
@@ -1,9 +1,9 @@
-const { Order } = require('../models');
+const { Orders } = require('../models');
 const MapperResponse = require('../utils/formatResponse');
 class OrderDao {
   static async createOrder(body) {
     try {
-      const data = await Order.create(body);
+      const data = await Orders.create(body);
       const format = MapperResponse.cleanDataDb(data);
       return format;
     } catch (error) {
